+++ examples/github-repository-webhook/main.tf
@@ -14,10 +14,10 @@ module "github_repository_webhook" {
   github_token        = var.github_token
   github_organization = var.github_organization
 
-  // Fetching these attributes from created already Atlantis Terraform state file
-  //
-  // This assumes that you are the owner of these repositories and they are available at:
-  // https://github.com/mygithubusername/awesome-repo and https://github.com/mygithubusername/another-awesome-repo
+  # Fetching these attributes from created already Atlantis Terraform state file
+  #
+  # This assumes that you are the owner of these repositories and they are available at:
+  # https://github.com/mygithubusername/awesome-repo and https://github.com/mygithubusername/another-awesome-repo
   atlantis_allowed_repo_names = data.terraform_remote_state.atlantis.outputs.atlantis_allowed_repo_names
 
   webhook_url    = data.terraform_remote_state.atlantis.outputs.atlantis_url_events