+++ modules/github-repository-webhook/versions.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_version = ">= 0.12.26"
+
+  required_providers {
+    github = {
+      source  = "integrations/github"
+      version = ">= 2.4.1"
+    }
+  }
+}