+++ modules/gitlab-repository-webhook/versions.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_version = ">= 0.13"
+
+  required_providers {
+    gitlab = {
+      source  = "gitlabhq/gitlab"
+      version = ">= 3.0"
+    }
+  }
+}